+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  * Added support for `*flush-on-newline*` to flush the `prn` and `println` output stream after the last newline (#865)
  * Added support for binding destructuring in `for` bindings (#774)
  * Added `==` as an alias to `=` (#859)
+ * Added custom exception formatting for `basilisp.lang.compiler.exception.CompilerException` and `basilisp.lang.reader.SyntaxError` to show more useful details to users on errors (#???)
 
 ### Changed
  * Cause exceptions arising from compilation issues during macroexpansion will no longer be nested for each level of macroexpansion (#852)